gcp/requirements.txt
tensorflow==2.10
google-cloud-storage
Pillow